+++ src/data/GPSData.java
@@ -7,8 +7,8 @@ import Sensors.GPS;
 @XmlRootElement
 public class GPSData {
 	private long id;
-	private long latitude;
-	private long longitude;
+	private double latitude;
+	private double longitude;
 
 	public GPSData() {
 	}
@@ -21,19 +21,19 @@ public class GPSData {
 		this.id = id;
 	}
 
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
